+++ index.html
@@ -6,6 +6,9 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script type="module" defer src="js/main.js"></script>
+  <script type="module" defer src="js/constants.js"></script>
+  <script type="module" defer src="js/util.js"></script>
+  <script type="module" defer src="js/data.js"></script>
   <title>Кексобукинг</title>
 </head>
 <body>
